1. Methods `same` and `sameCount` moved to namespace `Summary`. 2. Tests moved to directory `Summary`. 3. README updated.
